adding caikit-style prompt adapter support
