+++ vllm/prompt_adapter/models.py
@@ -6,10 +6,11 @@ import torch
 from peft.utils import load_peft_weights
 from torch import nn
 
-from vllm.adapter_commons.models import (AdapterLRUCache, AdapterModel,
+from vllm.adapter_commons.models import (AdapterLRUCache,
                                          AdapterModelManager)
 from vllm.config import PromptAdapterConfig
 from vllm.prompt_adapter.layers import PromptAdapterMapping
+from vllm.tgis_utils.prompt_adapter import PromptAdapterModel
 
 logger = logging.getLogger(__name__)
 
@@ -83,30 +84,6 @@ def convert_mapping(
         sampler_indices_padded.shape[-1]
     ]
     return (base_indices, sampler_indices, sampler_indices_padded, indices_len)
-
-
-class PromptAdapterModel(AdapterModel):
-
-    def __init__(self,
-                 prompt_adapter_id=None,
-                 num_virtual_tokens=None,
-                 prompt_embedding=None) -> None:
-        self.id = prompt_adapter_id
-        self.kv_cache = None
-        self.prompt_embedding = prompt_embedding
-        self.num_virtual_tokens = num_virtual_tokens
-
-    @classmethod
-    def from_local_checkpoint(cls,
-                              adapter_model_and_path,
-                              prompt_adapter_id,
-                              torch_device='cuda') -> "PromptAdapterModel":
-        adapters_weights = load_peft_weights(adapter_model_and_path,
-                                             torch_device)
-        prompt_embedding = adapters_weights["prompt_embeddings"].half()
-        num_virtual_tokens = prompt_embedding.shape[0]
-        return cls(prompt_adapter_id, num_virtual_tokens, prompt_embedding)
-
 
 class PromptAdapterModelManager(AdapterModelManager):
     """A manager that manages multiple Prompt Adapter models."""
